pyproject.toml
[tool.poetry]
name = "scrapnhs"
version = "0.1.0"
description = ""
authors = ["AxelTheRabbit <pietroconvalle@python.it>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
openrouteservice = "^2.3.3"
load-dotenv = "^0.1.0"
python-telegram-bot = "^21.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
